requirements.txt
anyio==3.6.1
argcomplete==2.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
Babel==2.10.3
backcall==0.2.0
Beaker==1.10.0
beautifulsoup4==4.11.0
bleach==5.0.1
blivet==3.4.4
blivet-gui==2.3.0
Brlapi==0.8.4
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.11
chrome-gnome-shell==0.0.0
click==8.0.4
cryptography==36.0.0
cupshelpers==1.0
dasbus==1.6
dbus-python==1.2.18
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
distro==1.6.0
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.1
fedora-third-party==0.10
fros==1.1
gpg==1.15.1
humanize==3.13.1
idna==3.3
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
json5==0.9.8
jsonschema==4.7.2
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyterlab==3.4.3
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.0
langtable==0.0.58
libcomps==0.1.18
lxml==4.7.1
Mako==1.1.4
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclassic==0.4.3
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
nftables==0.1
notebook-shim==0.1.0
numpy==1.23.1
olefile==0.46
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
Paste==3.5.0
pexpect==4.8.0
pickleshare==0.7.5
pid==2.2.3
Pillow==9.1.0
ply==3.11
productmd==1.33
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.1
ptyprocess==0.6.0
pure-eval==0.2.2
pwquality==1.4.4
pyasn1==0.4.8
pycairo==1.21.0
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.15.0
pycups==2.0.1
pycurl==7.45.1
pydub==0.25.1
pyenchant==3.2.2
Pygments==2.12.0
PyGObject==3.42.1
pykickstart==3.36
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyparted==3.12.0
PyQt5==5.15.6
PyQt5-sip==12.9.0
pyrsistent==0.18.1
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-meh==0.50
pytz==2022.1
pyudev==0.22.0
pyxdg==0.27
pyzmq==23.2.0
regex==2022.6.2
requests==2.27.1
requests-file==1.5.1
requests-ftp==0.3.1
rpm==4.17.1
rsa==4.8
selinux==3.3
Send2Trash==1.8.0
sepolicy==3.3
setools==4.4.0
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sniffio==1.2.0
sos==4.3
soupsieve==2.3.1
stack-data==0.3.0
systemd-python==234
Tempita==0.5.2
terminado==0.15.0
tinycss2==1.1.1
tk==0.1.0
torbrowser-launcher==0.3.5
tornado==6.2
traitlets==5.3.0
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
